Reorganise app folder structure
