+++ src/modules/sem/components/HeroImage.jsx
